This adds a listener to custom event to update the map position not only with the geosearch field, but also with any field in nova.
Any nova component just has to emit an event like this:

Nova.$emit("newcenter",{lat,lng})

and the map will pan to the new position.

If you need to use a custom name for the event you can define it in your field like this:

			MapMarker::make(__('Location'))
				->latitude('latitud')
				->longitude('longitud')
				->defaultLatitude(config('map.lat'))
				->defaultLongitude(	config('map.long'))
				->defaultZoom(config('map.zoom'))
				->centerCircle(config('map.radius'),config('map.color'),config('map.border'),config('map.opacity'))
				->listenToEventName('myCustomEventName')
				->required()->hideFromIndex()

If you don't specify a custom event name, "newcenter" will be used by default.